refactor: change login page structure of code
