+++ index.html
@@ -19,5 +19,6 @@
   </head>
   <body>
     <div id="root"></div>
+    <script src="./src/index.tsx" type="module"></script>
   </body>
 </html>
