+++ src/app/app.tsx
@@ -1,0 +1,6 @@
+import type { ReactNode } from "react";
+import { LoginPage } from "../pages";
+
+export function App(): ReactNode {
+  return <LoginPage />;
+}
