+++ src/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { App } from "./app";
+
+const root = document.querySelector("#root");
+
+ReactDOM.createRoot(root).render(<App />);
