+++ src/pages/login/ui/form/index.ts
@@ -1,0 +1,1 @@
+export { Form } from "./form";
